==> scripts/02_figures.Rmd <==
---
title: "02_figures"
author: "Adriana Uscanga"
date: '2022-05-16'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Figures 1 and 2. Timeline and Study Area Map

```{r}
# Load packages and datasets for making figures:

library(tidyverse)
library(lubridate)
library(ggplot2)
library(ggpubr)
library(cowplot)
library(rlang)
library(ggsignif)
library(ggh4x)

load("output/plots_cf_rs.RData")
load("output/rs_shorter_4p.RData")
```

# Figure 3. Break example

```{r}
# load p66139_2 data set
p66139_2 <- read_csv("output/ts_plots/66139_2.csv")


annual <- p66139_2 %>%
  arrange(sat_time) %>%
  mutate(year = year(date)) %>%
  group_by(year) %>%
  summarize(sat_time = as.numeric(sat_time),
            time1= paste0(year,"-01-01"),
            time2 = paste0(year+1,"-01-01"),
            ndvi_annual_sd = sd(ndvi),
            ndvi_annual_mean = mean(ndvi),
            ndvi_sd_1 = ndvi_annual_mean - ndvi_annual_sd,
            ndvi_sd_2 = ndvi_annual_mean + ndvi_annual_sd,
            ndvi_annual_min = min(ndvi),
            is_min = ifelse(ndvi == ndvi_annual_min, "min", NA)) %>%
  mutate(time1 = as_date(time1),
         time2 = as_date(time2))

p66139_2 <- p66139_2 %>%
  left_join(annual, by = "sat_time")

mean(p66139_2$ndvi)
sd(p66139_2$ndvi)
min(p66139_2$ndvi)
```


```{r}
#Panel A:
  
p66139_2 %>%
  mutate(year = year(date),
         date = as_date(date),
         is_min = ifelse(is.na(is_min), "no min", "min")) %>%
  ggplot(aes(x= date, y = ndvi)) +
  scale_x_date(date_breaks = "1 year",
               date_labels = "%Y") +
  # annual lines
  geom_vline(xintercept = as.numeric(p66139_2$time1[c(1,16,31,43,56,63,80,95,120,132,138,143,151,156,160,167,171,181,192,202,210,217,226,233,240,250,258,270)]),
             color = "gray90",
             linetype = 5,
             size = 0.25) +
  geom_vline(xintercept = as.numeric(p66139_2$time1[c(270)]+365),
             color = "gray90",
             linetype = 5,
             size = 0.25) +
  # ndvi values - time series
  geom_line(aes(x= date, y = ndvi),
            color = "grey65",
            alpha = 0.75) +
  scale_y_continuous(limits = c(0.25,1)) +
  # ndvi values - time series
  geom_point(aes(x= date, y = ndvi),
            color = "grey65",
            alpha = 0.75) +
  # breaks
  geom_vline(xintercept = as.numeric(p66139_2$date[c(71, 119, 260)]),
             color = "gray20",
             linetype = 2,
             size = 0.75) +
  # box for panel B
  geom_rect(aes(xmin = p66139_2$time1[250]-50,
                xmax = p66139_2$time2[270]+50,
                ymin = 0.55, ymax = 0.92),
            fill = NA,
            color = "gray25",
            size = 0.75) +
  theme_classic2() +
  labs(x= "Time (years)", y = "NDVI") +
  theme(axis.text = element_text(size = 8)) +
  theme(axis.title.x = element_text(size = 10)) +
  theme(axis.title.y = element_text(size = 10))

#Panel B:
p66139_2 %>%
  mutate(year = year(date),
         date = as_date(date),
         is_min = ifelse(is.na(is_min), "no min", "min")) %>%
  filter(date > as.numeric(p66139_2$date[255])) %>%
  ggplot(aes(x= date, y = ndvi)) +
  scale_x_date(date_breaks = "1 year",
               date_labels = "%Y") +
  # annual lines
  geom_vline(xintercept = as.numeric(p66139_2$time1[c(1,16,31,43,56,63,80,95,120,132,138,143,151,156,160,167,171,181,192,202,210,217,226,233,240,250,258,270)]),
             color = "gray90",
             linetype = 5,
             size = 0.25) +
  geom_vline(xintercept = as.numeric(p66139_2$time1[c(270)]+365),
             color = "gray90",
             linetype = 5,
             size = 0.25) +
  # ndvi values - time series
  geom_line(aes(x= date, y = ndvi),
            color = "grey65",
            alpha = 0.75) +
  scale_y_continuous(limits = c(0.55,0.9)) +
  # ndvi values - time series
  geom_point(aes(x= date, y = ndvi),
            color = "grey65",
            alpha = 0.75) +
  # breaks
  geom_vline(xintercept = as.numeric(p66139_2$date[c(71, 119, 260)]),
             color = "gray20",
             linetype = 2,
             size = 0.75) +
  #   # mean NDVI (full time series)
  # geom_hline(yintercept = 0.7632497,
  #            color = "#66c2a5",
  #            size = 1,
  #            linetype = 3) +
  #   # NDVI sd (full time series)
  # geom_hline(aes(yintercept= 0.7632497 + 0.08261055),
  #                color = "#66c2a5",
  #                size = 0.5,
  #                linetype = 3) +
  # geom_hline(aes(yintercept= 0.7632497 - 0.08261055),
  #                color = "#66c2a5",
  #                size = 0.5,
  #                linetype = 3) +
  # landsat images
  geom_point(aes(x= p66139_2$date[258],
                 y = p66139_2$ndvi[258]),
             shape = 5,
             size = 3,
             color = "orange1") +
  geom_point(aes(x= p66139_2$date[260],
                 y = p66139_2$ndvi[260]),
             shape = 5,
             size = 3,
             color = "orange1") +
  geom_point(aes(x= p66139_2$date[263],
                 y = p66139_2$ndvi[263]),
             shape = 5,
             size = 3,
             color = "orange1") +
  geom_point(aes(x= p66139_2$date[265],
                 y = p66139_2$ndvi[265]),
             shape = 5,
             size = 3,
             color = "orange1") +
  # planet images
  geom_point(aes(x= p66139_2$date[258]+23,
                 y = 0.83),
             shape = 4,
             size = 3,
             color = "purple3") +
  geom_point(aes(x= p66139_2$date[260],
                 y = p66139_2$ndvi[260]),
             shape = 4,
             size = 3,
             color = "purple3") +
  geom_point(aes(x= p66139_2$date[262]-4,
                 y = 0.65),
             shape = 4,
             size = 3,
             color = "purple3") +
  geom_point(aes(x= p66139_2$date[266]+1,
                 y = 0.805),
             shape = 4,
             size = 3,
             color = "purple3") +
  theme_classic2() +
  labs(x= "Time (years)", y = "NDVI") +
  theme(axis.text = element_text(size = 8)) +
  theme(axis.title.x = element_text(size = 10)) +
  theme(axis.title.y = element_text(size = 10))

#possible planet images:
#aug 27, may 18, april 20, feb 15
#,260,263,265)]
# ,260,263,265)])

```


# Figure 4. Variables
```{r}


```

Panel A:

```{r}

p66139_2 %>%
  mutate(year = year(date),
         date = as_date(date),
         is_min = ifelse(is.na(is_min), "no min", "min")) %>%
  ggplot(aes(x= date, y = ndvi)) +
  scale_x_date(date_breaks = "1 year",
               date_labels = "%Y") +
  # annual lines
  geom_vline(xintercept = as.numeric(p66139_2$time1[c(1,16,31,43,56,63,80,95,120,132,138,143,151,156,160,167,171,181,192,202,210,217,226,233,240,250,258,270)]),
             color = "gray90",
             linetype = 5,
             size = 0.25) +
  geom_vline(xintercept = as.numeric(p66139_2$time1[c(270)]+365),
             color = "gray90",
             linetype = 5,
             size = 0.25) +
  # annual boxes
  # geom_rect(aes(xmin = time1, xmax = time2, ymin = ndvi_sd_1, ymax = ndvi_sd_2),
  #           fill = "gray95",
  #           #color= "gray80",
  #           alpha = 0.10) +
  # annual ndvi mean
  # geom_segment(aes(x = time1, xend = time2, y = ndvi_annual_mean, yend = ndvi_annual_mean),
  #              color = "#66c2a5",
  #              size = 0.50,
  #              linetype = 1,
  #              alpha = 0.25) +
  # geom_point(aes(x= time2 - 182.5, y = ndvi_annual_mean),
  #            color = "#66c2a5",
  #            shape = 16) +
  # geom_errorbar(aes(x= time2 - 182.5,
  #                   ymin = ndvi_annual_mean - ndvi_annual_sd,
  #                   ymax= ndvi_annual_mean + ndvi_annual_sd),
  #               color = "#66c2a5") +
  # annual ndvi sd
  # geom_segment(aes(x = time1, xend = time2, y = ndvi_sd_1, yend = ndvi_sd_1),
  #              color = "#66c2a5",
  #              #size = 1,
  #              linetype = 3) +
  # geom_segment(aes(x = time1, xend = time2, y = ndvi_sd_2, yend = ndvi_sd_2),
  #              color = "#66c2a5",
  #              #size = 1,
  #              linetype = 3) +
 
  # ndvi values - time series
  geom_line(aes(x= date, y = ndvi),
            color = "grey65",
            alpha = 0.75) +
  scale_y_continuous(limits = c(0.25,1)) +
  # ndvi values - time series
  geom_point(aes(x= date, y = ndvi, shape = is_min, color = is_min, fill = is_min),
             alpha = 0.90) +
  scale_shape_manual(values = c(25,16)) +
  scale_fill_manual(values = c("#fc8d62", "gray80")) +
  scale_color_manual(values = c("#fc8d62", "gray80")) +
  # data collection date
  geom_vline(xintercept = as.numeric(p66139_2$date[203]),
             color = "gray60",
             linetype = "solid",
             size = 0.75) +
  # mean NDVI (full time series)
  geom_hline(yintercept = 0.7632497,
             color = "#66c2a5",
             size = 1,
             linetype = 3) +
    # NDVI sd (full time series)
  geom_hline(aes(yintercept= 0.7632497 + 0.08261055),
                 color = "#66c2a5",
                 size = 0.5,
                 linetype = 3) +
  geom_hline(aes(yintercept= 0.7632497 - 0.08261055),
                 color = "#66c2a5",
                 size = 0.5,
                 linetype = 3) +
  # min ndvi value
  geom_point(aes(x= p66139_2$date[73], y = 0.2649842),
             shape = 6,
             size= 4,
             color = "#fc8d62") +
  # age since last disturbance
   annotate("segment",
            x = p66139_2$date[119]+10,
            xend = p66139_2$date[203]-10,
            y = 0.4,
            yend = 0.4,
            color = "#8da0cb",
            size = 0.75,
            arrow = arrow(ends = "both", length = unit(.15,"cm"))) +
  # breaks
  geom_vline(xintercept = as.numeric(p66139_2$date[c(71, 119, 260)]),
             color = "gray20",
             linetype = 2,
             size = 0.75) +
  # box for panel B
  geom_rect(aes(xmin = p66139_2$time1[1]-30,
                xmax = p66139_2$time1[43]+30,
                ymin = 0.55, ymax = 0.9),
            fill = NA,
            color = "gray25",
            size = 0.75) +
  # box for panel C
  geom_rect(aes(xmin = p66139_2$date[255],
                xmax = p66139_2$date[270],
                ymin = 0.55, ymax = 0.9),
            fill = NA,
            color = "gray25",
            size = 0.75) +
# geom_segment(aes(x = p66139_2$date[119],
#                    xend = p66139_2$date[203],
#                    y = 0.4,
#                    yend = 0.4),
#                color = "#8da0cb",
#                size = 1,
#                linetype = 1,
#                lineend = "round") +
  theme_classic2() +
  theme(legend.position = "none") +
  labs(x= "Time (years)", y = "NDVI") +
  theme(axis.text = element_text(size = 8)) +
  theme(axis.title.x = element_text(size = 10)) +
  theme(axis.title.y = element_text(size = 10))

```

Panel B:

```{r}
p66139_2 %>%
  mutate(year = year(date),
         date = as_date(date),
         is_min = ifelse(is.na(is_min), "no min", "min")) %>%
  filter(year < 1996) %>%
  # base
  ggplot(aes(x= date, y = ndvi)) +
  scale_x_date(date_breaks = "1 year",
               date_labels = "%Y") +
  # annual lines
  geom_vline(xintercept = as.numeric(p66139_2$time1[c(1,16,31,43)]),
             color = "gray90",
             linetype = 5,
             size = 0.25) +
  # annual boxes
  geom_rect(aes(xmin = time1, xmax = time2, ymin = ndvi_sd_1, ymax = ndvi_sd_2),
            fill = "gray95",
            #color= "gray80",
            alpha = 0.10) +
    # time series
  geom_line(aes(x= date, y = ndvi),
            color = "grey65",
            alpha = 0.75) +
   # annual ndvi sd
  # geom_segment(aes(x = time1, xend = time2, y = ndvi_sd_1, yend = ndvi_sd_1),
  #              color = "#66c2a5",
  #              size = 0.75,
  #              linetype = 1) +
  # geom_segment(aes(x = time1, xend = time2, y = ndvi_sd_2, yend = ndvi_sd_2),
  #              color = "#66c2a5",
  #              size = 0.75,
  #              linetype = 1) +
  
  # ndvi values - time series
  geom_point(aes(x= date, y = ndvi, shape = is_min, color = is_min, fill = is_min),
             alpha = 0.90) +
  scale_shape_manual(values = c(25,16)) +
  scale_fill_manual(values = c("#fc8d62", "gray80")) +
  scale_color_manual(values = c("#fc8d62", "gray80")) +
   # annual ndvi mean
  geom_segment(aes(x = time1, xend = time2, y = ndvi_annual_mean, yend = ndvi_annual_mean),
               color = "gray60",
               size = 0.75,
               linetype = 1) +
  # geom_point(aes(x= time2 - 182.5, y = ndvi_annual_mean),
  #            color = "#66c2a5",
  #            shape = 16) +
  # geom_errorbar(aes(x= time2 - 182.5,
  #                   ymin = ndvi_annual_mean - ndvi_annual_sd,
  #                   ymax= ndvi_annual_mean + ndvi_annual_sd),
  #               color = "#66c2a5") +
  geom_hline(yintercept = 0.7632497,
             color = "#66c2a5",
             size = 0.75,
             linetype = 3) +
  geom_hline(aes(yintercept= 0.7632497 + 0.08261055),
                 color = "#66c2a5",
                 size = 0.5,
                 linetype = 3) +
  geom_hline(aes(yintercept= 0.7632497 - 0.08261055),
                 color = "#66c2a5",
                 size = 0.5,
                 linetype = 3) +
  theme_classic2() +
  theme(legend.position = "none") +
  labs(x= "Time (years)", y = "NDVI") +
  scale_y_continuous(limits = c(0.5,0.95))

```


# Figure 5. Forest structure ~ RS variables

```{r}
# load rs_sites_4p data set

load("output/rs_sites_4p.RData")

```


```{r}

# Basal area

f5_ba <- rs_sites_4p %>%
  mutate("Elevation (m asl)" = as.numeric(altitude),
         "NDVIa S.D." = as.numeric(ndvi_annual_sd),
         "NDVI S.D." = as.numeric(ndvi_sd_ts),
         "NDWIa min" = as.numeric(ndwi_annual_min),
         "Slope (degrees)" = as.numeric(slope_gee)) %>%
  pivot_longer(c(-site, -no_plots, -tree_density, -loreys_height, -basal_area, -agb, -is_unexpected ), names_to = "covariates", values_to = "value") %>%
  filter(covariates == "Elevation (m asl)" | 
         covariates == "NDVIa S.D." |
         covariates == "NDVI S.D." |
         covariates == "NDWIa min" |
         covariates == "Slope (degrees)") %>%
  ggplot(aes(value, basal_area)) +
  geom_point() +
  geom_smooth(method = "lm", color = "gray30", alpha = 0.15) +
  theme_bw() +
  facet_wrap(~covariates,
             scales = "free",
             ncol = 5) +
  xlab(NULL) +
  labs(y = (bquote('Basal Area ('*"m"~ha^-1*')')))

# Tree height

f5_lh <- rs_sites_4p %>%
  mutate("Elevation (m asl)" = as.numeric(altitude),
         "NDVIa S.D." = as.numeric(ndvi_annual_sd),
         "NDVI S.D." = as.numeric(ndvi_sd_ts),
         "NDWIa min" = as.numeric(ndwi_annual_min),
         "Slope (degrees)" = as.numeric(slope_gee)) %>%
  pivot_longer(c(-site, -no_plots, -tree_density, -loreys_height, -basal_area, -agb, -is_unexpected ), names_to = "covariates", values_to = "value") %>%
  filter(covariates == "Elevation (m asl)" | 
         covariates == "NDVIa S.D." |
         covariates == "NDVI S.D." |
         covariates == "NDWIa min" |
         covariates == "Slope (degrees)") %>%
  ggplot(aes(value, loreys_height)) +
  geom_point() +
  geom_smooth(method = "lm", color = "gray30", alpha = 0.15) +
  theme_bw() +
  facet_wrap(~covariates,
             scales = "free",
             ncol = 5) +
  xlab(NULL) +
  labs(y = "Lorey's height (m)")

# Tree density

f5_td <- rs_sites_4p %>%
  mutate("Elevation (m asl)" = as.numeric(altitude),
         "NDVIa S.D." = as.numeric(ndvi_annual_sd),
         "NDVI S.D." = as.numeric(ndvi_sd_ts),
         "NDWIa min" = as.numeric(ndwi_annual_min),
         "Slope (degrees)" = as.numeric(slope_gee)) %>%
  pivot_longer(c(-site, -no_plots, -tree_density, -loreys_height, -basal_area, -agb, -is_unexpected ), names_to = "covariates", values_to = "value") %>%
  filter(covariates == "Elevation (m asl)" | 
         covariates == "NDVIa S.D." |
         covariates == "NDVI S.D." |
         covariates == "NDWIa min" |
         covariates == "Slope (degrees)") %>%
  ggplot(aes(value, tree_density)) +
  geom_point() +
  geom_smooth(method = "lm", color = "gray30", alpha = 0.15) +
  theme_bw() +
  facet_wrap(~covariates,
             scales = "free",
             ncol = 5) +
  xlab(NULL) +
  labs(y = (bquote('Tree density ('*"tree"~ha^-1*')')))

# AGB

figure5_t <- rs_sites_4p %>%
  mutate("Elevation (m asl)" = as.numeric(altitude),
         "NDVIa S.D." = as.numeric(ndvi_annual_sd),
         "NDVI S.D." = as.numeric(ndvi_sd_ts),
         "NDWIa min" = as.numeric(ndwi_annual_min),
         "Slope (degrees)" = as.numeric(slope_gee)) %>%
  pivot_longer(c(-site, -no_plots, -tree_density, -loreys_height, -basal_area, -agb, -is_unexpected ),
               names_to = "covariates",
               values_to = "value") %>%
  mutate("log AGB" = log1p(agb),
         "sqrt Basal area" = sqrt(basal_area),
         "sqrt Lorey's height" = sqrt(loreys_height),
         "log Tree density" = log1p(tree_density)) %>%
  pivot_longer(c("log AGB", "sqrt Basal area", "sqrt Lorey's height", "log Tree density"),
               names_to = "forest_str",
               values_to = "f_str_values") %>%
  filter(covariates == "Elevation (m asl)" | 
         covariates == "NDVIa S.D." |
         covariates == "NDVI S.D." |
         covariates == "NDWIa min" |
         covariates == "Slope (degrees)") %>%
  ggplot(aes(value, f_str_values)) +
  geom_point(color = "gray20",
             alpha= 0.75) +
  geom_smooth(method = "lm",
              color = "gray10",
              se=F) +
  theme_bw() +
  theme(axis.text = element_text(size = 8)) +
  facet_grid(forest_str ~ covariates,
             scales = "free") +
  xlab(NULL) +
  ylab(NULL) 

figure5 <- rs_sites_4p %>%
  mutate("Elevation (m asl)" = as.numeric(altitude),
         "NDVIa S.D." = as.numeric(ndvi_annual_sd),
         "NDVI S.D." = as.numeric(ndvi_sd_ts),
         "NDWIa min" = as.numeric(ndwi_annual_min),
         "Slope (degrees)" = as.numeric(slope_gee)) %>%
  pivot_longer(c(-site, -no_plots, -tree_density, -loreys_height, -basal_area, -agb, -is_unexpected ),
               names_to = "covariates",
               values_to = "value") %>%
  mutate("AGB" = as.numeric(agb),
  "Basal area" = as.numeric(basal_area),
  "Lorey's height" = as.numeric(loreys_height),
  "Tree density" = as.numeric(tree_density)) %>%
  pivot_longer(c("AGB", "Basal area", "Lorey's height", "Tree density"),
               names_to = "forest_str",
               values_to = "f_str_values") %>%
  filter(covariates == "Elevation (m asl)" | 
         covariates == "NDVIa S.D." |
         covariates == "NDVI S.D." |
         covariates == "NDWIa min" |
         covariates == "Slope (degrees)") %>%
  ggplot(aes(value, f_str_values)) +
  geom_point(color = "gray30",
             alpha= 0.75) +
  # geom_smooth(method = "lm",
  #             color = "gray10",
  #             se=F) +
  theme_bw() +
  theme(axis.text = element_text(size = 8)) +
  facet_grid(forest_str ~ covariates,
             scales = "free") +
  xlab(NULL) +
  ylab(NULL)

library(ggh4x)

figure5 <- figure5 + facetted_pos_scales(
  y = list(
    forest_str == "AGB" ~ scale_y_continuous(trans = "log10"),
    forest_str == "Basal area" ~ scale_y_continuous(trans = "sqrt"),
    forest_str == "Lorey's height" ~ scale_y_continuous(trans = "sqrt"),
    forest_str == "Tree density" ~ scale_y_continuous(trans = "sqrt")
  )
)

figure5 <- figure5 + geom_smooth(method = "lm",
                                 se = F,
                                 color = "gray10")

figure5 <- rs_sites_4p %>%
  mutate("Elevation (m asl)" = as.numeric(altitude),
         "NDVIa S.D." = as.numeric(ndvi_annual_sd),
         "NDVI S.D." = as.numeric(ndvi_sd_ts),
         "NDWIa min" = as.numeric(ndwi_annual_min),
         "Slope (degrees)" = as.numeric(slope_gee)) %>%
  pivot_longer(c(-site, -no_plots, -tree_density, -loreys_height, -basal_area, -agb, -is_unexpected ),
               names_to = "covariates",
               values_to = "value") %>%
  mutate("AGB (Mg/ha)" = log1p(agb),
  "Basal area (m/ha)" = sqrt(basal_area),
  "Lorey's height (m)" = sqrt(loreys_height),
  "Tree density (tree/ha)" = sqrt(tree_density)) %>%
  pivot_longer(c("AGB (Mg/ha)", "Basal area (m/ha)", "Lorey's height (m)", "Tree density (tree/ha)"),
               names_to = "forest_str",
               values_to = "f_str_values") %>%
  filter(covariates == "Elevation (m asl)" | 
         covariates == "NDVIa S.D." |
         covariates == "NDVI S.D." |
         covariates == "NDWIa min" |
         covariates == "Slope (degrees)") %>%
  ggplot(aes(value, f_str_values)) +
  geom_point(color = "gray40",
             alpha= 0.75,
             size = 1.5) +
   geom_smooth(method = "lm",
               color = "gray10",
               se=F) +
  theme_bw() +
  theme(axis.text = element_text(size = 8)) +
  facet_grid(forest_str ~ covariates,
             scales = "free") +
  xlab(NULL) +
  ylab(NULL)

library(ggh4x)


figure5 <- figure5 + facetted_pos_scales(
    y = list(
    forest_str == "AGB (Mg/ha)" ~ scale_y_continuous(breaks = c(0, 2.397895, 4.615121,5.303305, 5.993961,6.398595, 6.685861),
                                             labels = c("0", "10", "100", "200", "400","600", "800")),
    forest_str == "Basal area (m/ha)" ~ scale_y_continuous(breaks = c(0, 2.236068, 3.162278, 4.472136, 6.324555, 7.745967, 8.944272),
                                                    labels = c("0", "5", "10", "20", "40", "60", "80")),
    forest_str == "Lorey's height (m)" ~ scale_y_continuous(breaks = c(0, 2.236068, 3.162278, 3.872983, 4.472136, 5, 5.477226),
                                                        labels = c("0", "5", "10", "15", "20", "25", "30")),
    forest_str == "Tree density (tree/ha)" ~ scale_y_continuous(breaks = c(0, 10, 22.36068, 31.62278, 38.72983, 44.72136, 50),
                                                      labels = c("0", "100", "500", "1000", "1500", "2000", "2500"))
  )
)

ggsave("FIGURE5.png", figure5, width=18, height=14, units="cm", dpi = "print")

## Figure 5 with plots

figure5_plots_df <- rs_shorter_4p %>%
  mutate('"Elevation (m asl)"' = as.numeric(altitude),
         '"Slope (degrees)"' = as.numeric(slope),
         '"No. breaks"' = as.numeric(number_breaks),
         '"NDVIa S.D."' = as.numeric(ndvi_annual_sd),
         '"NDWIa min"' = as.numeric(ndwi_annual_min)) %>%
  pivot_longer(c(-site, -no_plots, -plot_id, -tree_density, -loreys_height, -basal_area_ha, -agb_plot_ha),
               names_to = "covariates",
               values_to = "value") %>%
  # mutate("AGB (Mg/ha)" = log1p(agb_plot_ha),
  #        "Basal area (m^2/ha)" = sqrt(basal_area_ha),
  #        "Lorey's height (m)" = sqrt(loreys_height),
  #        "Tree density (tree/ha)" = sqrt(tree_density)) %>%
  mutate(agb = log1p(agb_plot_ha),
         ba = sqrt(basal_area_ha),
         lh = sqrt(loreys_height),
         td = sqrt(tree_density)) %>%
  pivot_longer(c(agb, ba, lh, td),
               names_to = "forest_str",
               values_to = "f_str_values") %>%
  filter(covariates == '"Elevation (m asl)"' | 
         covariates == '"Slope (degrees)"' |
         covariates == '"No. breaks"' |
         covariates == '"NDVIa S.D."' |
         covariates == '"NDWIa min"')

figure5_plots_df$forest_str <- factor(figure5_plots_df$forest_str,
                                labels = c((bquote('AGB ('*"Mg"~ha^-1*')')),
                                           (bquote('Basal area ('*"m"^2 ~ha^-1*')')),
                                           (bquote('"Lorey´s height (m)"')),
                                           (bquote('Tree density ('*"tree"~ha^-1*')'))))


figure5_plots <- figure5_plots_df %>%
  ggplot(aes(value, f_str_values)) +
  geom_point(color = "gray50",
             alpha= 0.50,
             size = 1.5) +
   geom_smooth(method = "lm",
               color = "gray10",
               se=F) +
  theme_bw() +
  xlab(NULL) +
  ylab(NULL) +
  theme(axis.text = element_text(size = 8)) +
  facet_grid(forest_str ~ covariates,
             scales = "free", 
             labeller = label_parsed,
             switch = "both") +
  theme(strip.background = element_blank(),
        strip.placement.x = "bottom",
        strip.placement.y = "left")

figure5_plots <- figure5_plots + facetted_pos_scales(
    y = list(
    forest_str == (bquote('AGB ('*"Mg"~ha^-1*')')) ~ scale_y_continuous(breaks = c(0, 2.397895, 4.615121,5.303305, 5.993961,6.398595, 6.685861),
                                             labels = c("0", "10", "100", "200", "400","600", "800")),
    forest_str == (bquote('Basal area ('*"m"^2 ~ha^-1*')')) ~ scale_y_continuous(breaks = c(0, 2.236068, 3.162278, 4.472136, 6.324555, 7.745967, 8.944272),
                                                    labels = c("0", "5", "10", "20", "40", "60", "80")),
    forest_str == (bquote('"Lorey´s height (m)"')) ~ scale_y_continuous(breaks = c(0, 2.236068, 3.162278, 3.872983, 4.472136, 5, 5.477226),
                                                        labels = c("0", "5", "10", "15", "20", "25", "30")),
    forest_str == (bquote('Tree density ('*"tree"~ha^-1*')')) ~ scale_y_continuous(breaks = c(0, 10, 22.36068, 31.62278, 38.72983, 44.72136, 50),
                                                      labels = c("0", "100", "500", "1000", "1500", "2000", "2500"))
  )
)

ggsave("FIGURE5_revised.tiff", figure5_plots, width=19, height=15, units="cm", dpi = 600)
```

# Figure 6. Boxplots

```{r}
rs_sites_4p %>%
  mutate(is_break = ifelse(mean_breaks == 0, "no break", "break")) %>%
  pivot_longer(c(tree_density, loreys_height, basal_area, agb), names_to = "forest_str", values_to = "value") %>%
  ggplot(aes(x= is_break, y = value)) +
  geom_boxplot() +
  stat_compare_means(method = "t.test",
                     aes(label = ..p.signif..)) +
  theme_bw() +
  facet_wrap(~forest_str,
             scales = "free",
             ncol = 4)
  
load("output/plots_cf_rs.RData")

figure6b <- plots_cf_rs %>%
  mutate("log AGB" = log1p(agb_plot_ha),
         "AGB" = as.numeric(agb_plot_ha),
         "Basal area" = as.numeric(basal_area_ha),
         "sqrt Basal area" = sqrt(basal_area_ha),
         "sqrt Lorey's height" = sqrt(loreys_height),
         "Lorey's height" = as.numeric(loreys_height),
         "Tree density" = as.numeric(tree_density),
         "log Tree density" = log1p(tree_density)) %>%
  mutate(is_break = ifelse(number_breaks == 0, "no break", "break")) %>%
  pivot_longer(c("AGB", "Basal area", "Lorey's height", "Tree density"),
               names_to = "forest_str",
               values_to = "value") %>%
  ggplot(aes(x= is_break, y = value)) +
  geom_boxplot(fill = "gray80",
               outlier.alpha = 0.5) +
  # stat_compare_means(method = "t.test",
  #                    aes(label = ..p.signif..),
  #                    label.x = 1.4,
  #                    label.y = 1) +
  theme_bw() +
  xlab(NULL) +
  ylab(NULL) +
  facet_wrap(~forest_str,
             scales = "free",
             ncol = 4)

figure6b <- figure6b + facetted_pos_scales(
    y = list(
    forest_str == "AGB" ~ scale_y_continuous(trans = "log10"),
    forest_str == "Basal area" ~ scale_y_continuous(trans = "sqrt"),
    forest_str == "Lorey's height" ~ scale_y_continuous(trans = "sqrt"),
    forest_str == "Tree density" ~ scale_y_continuous(trans = "sqrt")
  )
) 

plots_cf_rs %>%
  mutate("log AGB" = log1p(agb_plot_ha),
         "AGB" = as.numeric(agb_plot_ha),
         "Basal area" = as.numeric(basal_area_ha),
         "Lorey's height" = as.numeric(loreys_height),
         "Tree density" = as.numeric(tree_density)) %>%
  mutate(is_break = ifelse(number_breaks == 0, "no break", "break")) %>%
  pivot_longer(c("log AGB", "Basal area", "Lorey's height", "Tree density"),
               names_to = "forest_str",
               values_to = "value") %>%
  ggplot(aes(x= tree_density)) +
  geom_density(aes(fill= is_break), alpha = 0.25)

plots_cf_rs %>%
  mutate(is_break = ifelse(number_breaks == 0, "no break", "break")) %>%
  group_by(is_break) %>%
  summarize(td = n(),
            td_mean = mean(tree_density))


figure6_df <- plots_cf_rs %>%
  mutate(agb = log1p(agb_plot_ha),
         #"AGB" = as.numeric(agb_plot_ha),
         #"Basal area" = as.numeric(basal_area_ha),
         ba = sqrt(basal_area_ha),
         lh = sqrt(loreys_height),
         #"Lorey's height" = as.numeric(loreys_height),
         #"Tree density" = as.numeric(tree_density),
         td = sqrt(tree_density)) %>%
  mutate(is_break = ifelse(number_breaks == 0, "no break", "break")) %>%
  pivot_longer(c(agb, ba, lh, td),
               names_to = "forest_str",
               values_to = "value")

figure6_df$forest_str <- factor(figure6_df$forest_str,
                                labels = c((bquote('AGB ('*"Mg"~ha^-1*')')),
                                           (bquote('Basal area ('*"m"^2 ~ha^-1*')')),
                                           (bquote('"Lorey´s height (m)"')),
                                           (bquote('Tree density ('*"tree"~ha^-1*')'))))
figure6 <- figure6_df %>%
  ggplot(aes(x= is_break, y = value)) +
  geom_boxplot(fill = "gray80",
               outlier.alpha = 0.5) +
  stat_compare_means(method = "t.test",
                     aes(label = ..p.signif..),
                     label.x = 1.4,
                     label.y.npc = "top") +
  theme_bw() +
  xlab(NULL) +
  ylab(NULL) +
  facet_wrap(~forest_str,
             scales = "free",
             ncol = 4,
             strip.position = "left", 
             labeller = label_parsed) +
  theme(strip.background = element_blank(),
        strip.placement = "outside")

figure6<- figure6 + facetted_pos_scales(
    y = list(
    forest_str == (bquote('AGB ('*"Mg"~ha^-1*')')) ~ scale_y_continuous(breaks = c(0, 2.397895, 4.615121,5.303305, 5.993961,6.398595, 6.685861),
                                             labels = c("1", "10", "100", "200", "400","600", "800")),
    forest_str == (bquote('Basal area ('*"m"^2 ~ha^-1*')')) ~ scale_y_continuous(breaks = c(0, 2.236068, 3.162278, 4.472136, 6.324555, 7.745967, 8.944272),
                                                    labels = c("0", "5", "10", "20", "40", "60", "80")),
    forest_str == (bquote('"Lorey´s height (m)"')) ~ scale_y_continuous(breaks = c(0, 2.236068, 3.162278, 3.872983, 4.472136, 5, 5.477226),
                                                        labels = c("0", "5", "10", "15", "20", "25", "30")),
    forest_str == (bquote('Tree density ('*"tree"~ha^-1*')')) ~ scale_y_continuous(breaks = c(0, 10, 22.36068, 31.62278, 38.72983, 44.72136, 50),
                                                      labels = c("0", "100", "500", "1000", "1500", "2000", "2500"))
  )
)

ggsave("FIGURE3_revised.tiff", figure6, width=18, height=12, units="cm", dpi = 600)


```

# Figure 7. Age

```{r}

figure7 <- plots_cf_rs %>%
  mutate("log AGB" = log1p(agb_plot_ha),
         "AGB" = as.numeric(agb_plot_ha),
         "Basal area" = as.numeric(basal_area_ha),
         "sqrt Basal area" = sqrt(basal_area_ha),
         "Lorey's height" = as.numeric(loreys_height),
         "Tree density" = as.numeric(tree_density),
         "sqrt Tree density" = sqrt(tree_density)) %>%
  mutate(is_break = ifelse(number_breaks == 0, "no break", "break")) %>%
  mutate(age_fixed = ifelse(number_breaks == 0, 22, age)) %>%
  pivot_longer(c("log AGB", "sqrt Basal area", "Lorey's height", "Tree density"),
               names_to = "forest_str",
               values_to = "value") %>%
  ggplot(aes(x= age, y = value)) +
  geom_point(aes(color = is_break),
             alpha = 0.50) +
  scale_color_manual(values = c("gray30", "gray85")) +
  geom_smooth(method = "lm",
              color = "gray10",
              se = F) +
  theme_bw() +
  ylab(NULL) +
  stat_regline_equation(aes(label = ..adj.rr.label..),
                        size = 3,
                        label.y.npc = 0.97) +
  theme(legend.title = element_blank()) +
  labs(x = "Time since last disturbance (years)") +
  facet_wrap(~forest_str,
             scales = "free",
             ncol = 4)

# BA p < 0.05 adj-R2 = 0.09
# AGB p < 0.05 adj-R2 = 0.13
# Lorey's height p < 0.05 adj-R2 = 0.08
# Tree density p < 0.002 adj-R2= 0.02


summary(lm(plots_cf_rs$tree_density ~ plots_cf_rs$age))

# figure7_db <- plots_cf_rs %>%
#   mutate("AGB (Mg/ha)" = log1p(agb_plot_ha),
#          "Basal area (m*m/ha)" = sqrt(basal_area_ha),
#          "Lorey's height (m)" = sqrt(loreys_height),
#          "Tree density (tree/ha)" = sqrt(tree_density)) %>%
#   mutate(is_break = ifelse(number_breaks == 0, "no break", "break")) %>%
#   filter(is_break == "break")

figure7_db <- plots_cf_rs %>%
  mutate(agb = log1p(agb_plot_ha),
         ba = sqrt(basal_area_ha),
         lh = sqrt(loreys_height),
         td = sqrt(tree_density)) %>%
  mutate(is_break = ifelse(number_breaks == 0, "no break", "break")) %>%
  filter(is_break == "break") %>%
  select(agb, ba, lh, td, age) %>%
  pivot_longer(c(agb, ba, lh, td),
               names_to = "forest_str",
               values_to = "value")

figure7_db$forest_str <- factor(figure7_db$forest_str,
                                labels = c((bquote('AGB ('*"Mg"~ha^-1*')')),
                                           (bquote('Basal area ('*"m"^2 ~ha^-1*')')),
                                           (bquote('"Lorey´s height (m)"')),
                                           (bquote('Tree density ('*"tree"~ha^-1*')'))))


figure7 <- figure7_db %>%
  # pivot_longer(c("AGB (Mg/ha)", "Basal area (m*m/ha)", "Lorey's height (m)", "Tree density (tree/ha)"),
  #              names_to = "forest_str",
  #              values_to = "value") %>%
  # pivot_longer(c(agb, ba, lh, td),
  #              names_to = "forest_str",
  #              values_to = "value") %>%
  ggplot(aes(x= age, y = value)) +
  geom_point(color = "gray70",
             size = 1.5) +
  geom_smooth(method = "lm",
              color = "gray10",
              se = F) +
  theme_bw() +
  ylab(NULL) +
  # stat_regline_equation(aes(label = ..adj.rr.label..),
  #                       size = 3,
  #                       label.y.npc = 0.03,
  #                       label.x.npc = 0.55) +
  theme(legend.title = element_blank()) +
  labs(x = "Time since last disturbance (years)") +
  facet_wrap(~forest_str,
             scales = "free",
             ncol = 4,
             strip.position = "left", 
             labeller = label_parsed) +
  theme(strip.background = element_blank(),
        strip.placement = "outside")

# (c("AGB (Mg/ha)" = (bquote('AGB ('*"Mg"~ha^-1*')')),
#                           "Basal area (m*m/ha)" = (bquote('Basal area ('*"m"^2 ~ha^-1*')')),
#                           "Lorey's height (m)" = "Lorey's height (m)",
#                           "Tree density (tree/ha)" = (bquote('Tree density ( '*"tree"~ha^-1*')'))))

# figure7 <- figure7 + facetted_pos_scales(
#     y = list(
#     forest_str == "AGB (Mg/ha)" ~ scale_y_continuous(breaks = c(0, 2.397895, 4.615121,5.303305, 5.993961,6.398595, 6.685861),
#                                              labels = c("0", "10", "100", "200", "400","600", "800")),
#     forest_str == "Basal area (m*m/ha)" ~ scale_y_continuous(breaks = c(0, 2.236068, 3.162278, 4.472136, 6.324555, 7.745967, 8.944272),
#                                                     labels = c("0", "5", "10", "20", "40", "60", "80")),
#     forest_str == "Lorey's height (m)" ~ scale_y_continuous(breaks = c(0, 2.236068, 3.162278, 3.872983, 4.472136, 5, 5.477226),
#                                                         labels = c("0", "5", "10", "15", "20", "25", "30")),
#     forest_str == "Tree density (tree/ha)" ~ scale_y_continuous(breaks = c(0, 10, 22.36068, 31.62278, 38.72983, 44.72136, 50),
#                                                       labels = c("0", "100", "500", "1000", "1500", "2000", "2500"))
#   )
# )

figure7 <- figure7 + facetted_pos_scales(
    y = list(
    forest_str == (bquote('AGB ('*"Mg"~ha^-1*')')) ~ scale_y_continuous(breaks = c(0,
                                                                                   2.397895,
                                                                                   4.615121,5.303305,
                                                                                   5.993961,
                                                                                   6.398595,
                                                                                   6.685861),
                                             labels = c("0",
                                                        "10",
                                                        "100",
                                                        "200",
                                                        "400",
                                                        "600",
                                                        "800")),
    forest_str == (bquote('Basal area ('*"m"^2 ~ha^-1*')')) ~ scale_y_continuous(breaks = c(0, 2.236068, 3.162278, 4.472136, 6.324555, 7.745967, 8.944272),
                                                    labels = c("0", "5", "10", "20", "40", "60", "80")),
    forest_str == (bquote('"Lorey´s height (m)"')) ~ scale_y_continuous(breaks = c(0, 2.236068, 3.162278, 3.872983, 4.472136, 5, 5.477226),
                                                        labels = c("0", "5", "10", "15", "20", "25", "30")),
    forest_str == (bquote('Tree density ('*"tree"~ha^-1*')')) ~ scale_y_continuous(breaks = c(0, 10, 22.36068, 31.62278, 38.72983, 44.72136, 50),
                                                      labels = c("0", "100", "500", "1000", "1500", "2000", "2500"))
  )
)

ggsave("FIGURE4_revised.tiff", figure7, width=18, height=4.5, units="cm", dpi = 600)

# BA p < 0.05 adj-R2 = 0.09
# AGB p < 0.05 adj-R2 = 0.13
# Lorey's height p < 0.05 adj-R2 = 0.08
# Tree density p < 0.002 adj-R2= 0.02


summary(lm(figure7_db$AGB ~ figure7_db$age))
summary(lm(figure7_db$`Basal area` ~ figure7_db$age))
summary(lm(figure7_db$`Tree density` ~ figure7_db$age))
summary(lm(figure7_db$`Lorey's height` ~ figure7_db$age))
```

# figure 7 in boxplot

```{r}
plots_cf_rs %>%
  mutate("log AGB" = log1p(agb_plot_ha),
         "AGB" = as.numeric(agb_plot_ha),
         "Basal area" = as.numeric(basal_area_ha),
         "Lorey's height" = as.numeric(loreys_height),
         "Tree density" = as.numeric(tree_density),
         "sqrt Tree density" = sqrt(tree_density)) %>%
  mutate(is_break = ifelse(number_breaks == 0, "no break", "break")) %>%
  mutate(age_fixed = ifelse(number_breaks == 0, 22, age)) %>%
  mutate(age_category = ifelse(age_fixed > 0 & age_fixed < 10, "0-10",
                               ifelse(age_fixed >= 10 & age_fixed < 20, "10-20",
                                      ">20"))) %>%
  pivot_longer(c("log AGB", "Basal area", "Lorey's height", "Tree density"),
               names_to = "forest_str",
               values_to = "value") %>%
  ggplot(aes(x= factor(age_category,
                       levels = c("0-10", "10-20", ">20"),
                       ordered = TRUE),
             y = value)) +
  geom_boxplot(fill = "gray80",
               outlier.alpha = 0.5) +
  stat_compare_means(method = "anova",
                     aes(label = ..p.signif..),
                     label.x = 2) +
  theme_bw() +
  ylab(NULL) +
  theme(legend.title = element_blank()) +
  labs(x = "Age category") +
  facet_wrap(~forest_str,
             scales = "free",
             ncol = 4)

aov_df <- plots_cf_rs %>%
  mutate(age_fixed = ifelse(number_breaks == 0, 22, age)) %>%
  mutate(age_category = ifelse(age_fixed >= 0 & age_fixed < 10, "0-10",
                               ifelse(age_fixed >= 10 & age_fixed < 20, "10-20",
                                      ">20"))) %>%
  select(age_category, agb_plot_ha, tree_density, basal_area_ha, loreys_height)

aov_df %>%
  group_by(age_category) %>%
  ggplot(aes(x = basal_area_ha)) +
  geom_histogram(aes(fill = age_category))

#  anova_test()

# BA anova:
aov_ba <- aov(aov_df$basal_area_ha ~ aov_df$age_category, data = aov_df)
summary(aov_ba)
TukeyHSD(aov_ba)

# log AGB anova
aov_agb <- aov(log1p(agb_plot_ha) ~ age_category, data = aov_df)
summary(aov_agb)
TukeyHSD(aov_agb)

# Lorey's height
aov_lh <- aov(loreys_height ~ age_category, data = aov_df)
summary(aov_lh)
TukeyHSD(aov_lh)

# Tree density
aov_td <- aov(tree_density ~ age_category, data = aov_df)
summary(aov_td)
TukeyHSD(aov_td)

```
